% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_matrix.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{Title}
\usage{
create_distance_matrix(pairwise_dist, diag_val = 0)
}
\arguments{
\item{pairwise_dist}{a list of pairwise comparisons from create_pairwise_comparisons}

\item{diag_val}{value of the diagonal, defaults to 0 (0 distance)}
}
\value{
a distance matrix
}
\description{
Title
}
\examples{
matrix <- replicate(20, rnorm(10))
create_distance_matrix(create_pairwise_comparisons(matrix))
}

