% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pairwise_comparisons.R
\name{create_pairwise_comparisons}
\alias{create_pairwise_comparisons}
\title{Create pairwise comparisons by row or column}
\usage{
create_pairwise_comparisons(matrix, ncores = 1, by = "row")
}
\arguments{
\item{matrix}{a matrix or data.frame containing samples and observations}

\item{ncores}{number of cores to parallelize by. Defaults to single-core}

\item{by}{either "row" or "column", to specify if rows or columns are samples. Defaults to "row"}
}
\value{
a list of data.frames
}
\description{
Create pairwise comparisons by row or column
}
\examples{
matrix <- replicate(20, rnorm(10))
create_pairwise_comparisons(matrix, ncores = 1) #not allowed to have parallel examples
}

