% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceParallel.R
\docType{package}
\name{distanceParallel}
\alias{distanceParallel}
\alias{distanceParallel-package}
\title{Parallel distance matrix calculations}
\description{
dist(x) in base R only operates on one thread. Using the foreach package, we can now create distance matrixes in parallel.
}

