% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_cores.R
\name{divide_cores}
\alias{divide_cores}
\title{Title}
\usage{
divide_cores(total, ncores = 1)
}
\arguments{
\item{total}{a number}

\item{ncores}{number of cores. If unspecified, will use all cores on machine}
}
\value{
data frame for use in foreach
}
\description{
Title
}
\examples{
divide_cores(total = 1000)
}

