% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distance_matrix.R
\name{create_distance_matrix}
\alias{create_distance_matrix}
\title{Title}
\usage{
create_distance_matrix(pairwise, ncores = 1, diag_val = 0,
  distfun = function(x1, x2) sqrt(sum((x1 - x2)^2)))
}
\arguments{
\item{pairwise}{a list of pairwise comparisons from create_pairwise_comparisons}

\item{ncores}{number of cores to parallelize this, defaults to single-core}

\item{diag_val}{value of the diagonal, defaults to 0 (0 distance)}

\item{distfun}{a function used to calculate distance between any pairwise comparisons. Defaults to sqrt(sum((x1 - x2) ^ 2)) (euclidian distance)}
}
\value{
a distance matrix
}
\description{
Title
}
\examples{
matrix <- replicate(20, rnorm(10))
create_distance_matrix(create_pairwise_comparisons(matrix))
}

